Build changes, unfinished work on AIS
I decided to finish AIS, but it became apparent as I started to write tests and add checks that what I've implemented so far is wrong.
I've completely lost interest in this, but I hate to have this unfinished code sitting in my repository. I'm not sure yet whether I'll reimplement it just for the sake of finishing or take the hit to my pride and recruitment appeal and do something more productive.